refactor: collapse secondary projectile draw modules
