--- src/crimson/render/projectile_draw/secondary_type2.py
@@ -1,20 +1,0 @@
-from __future__ import annotations
-
-from .secondary_rocket_common import SecondaryRocketStyle, draw_secondary_rocket_style
-from .types import SecondaryProjectileDrawCtx
-
-_TYPE2_ROCKET_STYLE = SecondaryRocketStyle(
-    base_size=10.0,
-    glow_size=40.0,
-    glow_rgb=(1.0, 1.0, 1.0),
-    glow_alpha_mul=0.58,
-)
-
-
-def draw_secondary_type2(ctx: SecondaryProjectileDrawCtx) -> bool:
-    if int(ctx.proj_type) != 2:
-        return False
-    return draw_secondary_rocket_style(ctx, style=_TYPE2_ROCKET_STYLE)
-
-
-__all__ = ["draw_secondary_type2"]
